boot: fix loading failure on Qemu/KVM
Qemu/KVM couldn't load kernel.img. The reason of failure is changing
disk geometry for Bochs emulator. Previosly it worked due because
"BIOS" in Qemu tries to predict disk parameters, however sometimes this
algorithm fails and it's hard to understand how to work with disk
in this case. The solution for this problem is to set explicitly CHS of disk.