mmu: fix using bootloader segments
A segment are updated when setting corresponding segment register. It means
that using of lgdt is not enough to setup own segment. Due to missing this
essential step bolgenos-ng used segments setup by multiboot loader
for a long time. This commit fixes the problem by setting all data segments
and long-jump to correct code segment.